Improve documentation.

+++ cli/delete-chore.sh
@@ -2,19 +2,20 @@
 
 usage="$(basename "$0") CHORE [OPTIONS]
 
-Delete a chore session.
+Delete a chore completion.
 Arguments:
     CHORE          The name of the chore completion to delete.
 
 Options:
     -h, --help     Show this hep text and exit.
     --preview      Show the SQL command to be executed.
-	--quiet        Suppress output.
+    --quiet        Suppress output.
     -v, --verbose  Show the SQL commands as they are excuted."
 
 # Process options
 a=0
 o=0
+execute=1
 for arg in "$@"
 do
 	if [[ $arg == "-h" ]] || [[ $arg == "--help" ]]
@@ -25,6 +26,9 @@ do
 	then
 		arguments[$a]=$arg
 		((a++))
+	elif [[ $arg == "--preview" ]]
+	then
+		execute=0
 	else
 		options[$o]=$arg
 		((o++))
@@ -42,6 +46,12 @@ fi
 chore=${arguments[0]//\'/\\\'}
 
 # Invoke SQL
-chore_completion_id=$(chore-get-completion-id "$chore")
-sql="SET @c = $chore_completion_id;CALL delete_chore_completion($chore_completion_id)"
-chore-database "$sql" ${options[@]}
+if [[ $execute -eq 1 ]]
+then
+	chore_completion_id=$(chore-get-completion-id "$chore")
+	sql="SET @c = $chore_completion_id;CALL delete_chore_completion($chore_completion_id)"
+	chore-database "$sql" ${options[@]}
+else
+	chore-get-completion-id "$chore" --preview
+	echo "CALL delete_chore_completion(@c)"
+fi
